flatten package structure, deleted CVTSigner class (#31)
