+++ src/main/python/covata/delta/crypto.py
