--- src/main/python/covata/delta/crypto/__init__.py
@@ -1,28 +1,0 @@
-#   Copyright 2017 Covata Limited or its affiliates
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
-from __future__ import absolute_import
-
-from .cryptoservice import serialize_public_key, calculate_sha256hex, \
-    generate_private_key, generate_initialization_vector, generate_secret_key, \
-    encrypt, decrypt, encrypt_key_with_public_key, decrypt_with_private_key, \
-    deserialize_public_key
-from .keystore import FileSystemKeyStore
-
-__all__ = ["FileSystemKeyStore",
-           "serialize_public_key", "calculate_sha256hex",
-           "generate_private_key", "generate_initialization_vector",
-           "generate_secret_key", "encrypt", "decrypt",
-           "encrypt_key_with_public_key", "decrypt_with_private_key",
-           "deserialize_public_key"]
